requirements.txt
gunicorn==21.2.0
Flask==2.3.3
google-cloud-storage==2.11.0
requests==2.31.0
moviepy==1.0.3
